dac_panel: Workaround a random issue with the panel's name.
For some unknown reason, the panel's name is always set to be
the same as the string for "digital_filter_5" and there's no code
pointing out to a bug or anything that might explain this issue.

Since we don't use that filter, we can rename it to the match the
name we expect the panel to have, and that "fixes" the naming
issue.

Change-Id: I4e60b9799a948d7c25ddb294127514c3de12d897