dac_panel: Get the actual DAC status before applying settings.
Previously, it was possible to apply the panel settings into the
DAC when it was turned off (bypass mode), resulting in many i2c
errors which were completely ignored by the panel, making it so
that the settings weren't really applied. This commit solves that,
with the panel now only applying the changes if the DAC is
confirmed to be enabled and in HiFi mode.

We still have a bug in here though: If the user does not explicitly
follow the panel's toast telling them to stop all media playback,
the routes won't update properly, resulting in no audio once the
panel applies its settings and the DAC goes into HiFi mode...
however, this might end up being a constraint of using the AOSP hal,
as we might not have a way of getting the status of global media
playback, much less influence it to do what we need it to.

Change-Id: Ib1b4424f2e591f16530363f4774ad4816fcac2d6