dac_panel: Add checks to see if media is playing during DAC setup
Previously the panel had no way of knowing whether media was
playing in order to update audio routes for the DAC (we can't swap
them while media is running, or we get no audio output as a result).

Now it does, and also notifies the user if media playback isn't
stopped by the time it is going to commit its changes, scheduling
a retry afterwards.

Change-Id: Ie7dc44689fa4428f31347b47d40e8dc3c4e41d1e