dac_panel: Add a variable to track changes that require a pause
This one fixes case that's not very common, but can result in a
panel crash, which is when the user requests multiple "big" changes
that require media playback to be paused briefly in order to apply.

In those cases, the panel will now refuse to apply any of them
until the first one goes through.

Change-Id: I7f0c75ba3b6e000c65238d0482df7c1ab558b7f5